+++ src/ska_tmc_dishleafnode/commands/track_command.py
@@ -111,9 +111,10 @@ class Track(DishLNCommand):
             )
             return result_code, message
 
-        result_code, message = self.call_adapter_method(
-            "Dish Master", self.dish_master_adapter, "Track"
-        )
+        with self.component_manager.tango_operation_execution_lock:
+            result_code, message = self.call_adapter_method(
+                "Dish Master", self.dish_master_adapter, "Track"
+            )
 
         if result_code[0] == ResultCode.FAILED:
             return result_code[0], message[0]