+++ src/ska_tmc_dishleafnode/manager/event_receiver.py
@@ -85,9 +85,9 @@ class DishLNEventReceiver(EventReceiver):
                     stateless=True,
                 )
                 dish_dev_proxy.subscribe_event(
-                    "longRunningCommandStatus",
+                    "longRunningCommandResult",
                     tango.EventType.CHANGE_EVENT,
-                    self.handle_long_running_command_status,
+                    self.handle_long_running_command_result,
                     stateless=True,
                 )
             except Exception as exception:
@@ -185,9 +185,10 @@ class DishLNEventReceiver(EventReceiver):
             return
         new_value = event_flag.attr_value.value
         self._component_manager.achieved_pointing_data.put(new_value)
-        self._logger.debug(f"achievedPointing value is updated to {new_value}")
+        # self._logger.debug(f"achievedPointing value is updated to
+        #  {new_value}")
 
-    def handle_long_running_command_status(
+    def handle_long_running_command_result(
         self, event_data: tango.EventData
     ) -> None:
         """Method to handle and update the latest value of
@@ -207,7 +208,7 @@ class DishLNEventReceiver(EventReceiver):
             )
             return
         new_value = event_data.attr_value.value
-        self._component_manager.update_device_long_running_command_status(
+        self._component_manager.update_device_long_running_command_result(
             new_value
         )
         self._logger.info(f"long running command value updated to {new_value}")